refactor(fbuild-deploy): split esp32_native.rs into submodules to satisfy LOC gate
